+++ lib/okcollect/okhi_location_manager.dart
@@ -63,8 +63,10 @@ class _OkHiLocationManagerState extends State<OkHiLocationManager> {
       );
     }
     return PopScope(
-      onPopInvoked: (didPop) {
-        _handleWillPopScope();
+      onPopInvokedWithResult: (didPop, result) {
+        if (didPop) {
+          _handleWillPopScope();
+        }
       },
       child: WebViewWidget(controller: _controller!),
     );
@@ -238,7 +240,6 @@ class _OkHiLocationManagerState extends State<OkHiLocationManager> {
   _handleMessageReceived(JavaScriptMessage jsMessage) {
     final Map<String, dynamic> data = jsonDecode(jsMessage.message);
     final String message = data["message"];
-    debugPrint(message);
     switch (message) {
       case "location_created":
       case "location_updated":