+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.36
+
+- fix: Android 16KB page size compatibility
+- fix: In progress verification addresses syncing
+
 ## 1.1.34
 
 - fix: Reply already submitted exception on android when fetching user's location